pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "db-intelligence"
version = "0.1.0"
description = "Natural language to Postgres SQL agent"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "DB Intelligence" }
]
dependencies = [
  "fastapi==0.111.0",
  "uvicorn==0.30.1",
  "python-dotenv==1.0.1",
  "pydantic==2.7.4",
  "requests>=2.31.0,<3",
  "psycopg[binary]>=3.1,<3.3",
  "psycopg-pool>=3.1,<3.3",
  "sqlparse==0.5.0",
  "tenacity==9.0.0",
  "SQLAlchemy>=2.0,<3.0",
  "streamlit>=1.34,<2.0",
]

[project.scripts]
dbi-cli = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "pyproject.toml"]